Implement a bespoke SMT database backend
- It's the fastest and smallest of everything tested so far.
- It's also the least durable.

Moves most of the code out of the `sqlite` module into `tree`, so it can
be used by any number of backends. `sqlite` and the new `nih` backends
both use it.

There is some trait shenanigans going on for code reuse. Moved the
`render` method to the main `Smt` trait to limit surface area on exposed
trait methods. So that its internal recursive inner methods don't have
to be made public.

The `monotree` backends technically have a `render` method now, but it
remains unimplemented (I don't know how to walk the tree with the
available public API.).

Made a few QoL improvements in the main module, like combining the
`smt_demo` functions into one that takes a parameter to enable diagrams.

We've decided that we want to benchmark up to 1 million KV pairs in the
SMTs, allowing for failures on backends that can't handle it. 1M is
still not enabled until something is setup for "failure recovery".

That's about it. There are still some TODOs sprinkled around, like the
new `nih` backend lacking a way to load the DB from disk. Next I am
going to start on property testing to make sure the tree algorithm is
solid.