<M-C-F3>Update arbtests
Remove the BTreeMap, which was associating hashes with random values.
That isn't really necessary. We can use the hashes as values!

Make sure to insert the leftmost and rightmost hashes before getting the
proofs.

Add a new arbtest that checks proof siblings in a range narrower than
the full keyspace boundaries.