Use sqlite prepared statements and batch transaction
Performs around 2x better. Both changes independently are each 50% of
the total improvement.

It's possible to squeeze out a small percentage more, but it is not at
all worth the added code complexity. This is a good compromise.

(The hack in question returns `Statement<'_>`s from `prepare()`. But
doing so requires threading the prepared statements through all of the
insert and get function signatures.)