Smooth 2D charts
This was a lot harder than I thought it would be!

Still incomplete. The simulation needs to compare the difference between
no-nonce and nonce (currently only showing magnitude with no-nonce).

This also removes storing the proofs in files. But might want to bring
it back. Running the simulation is VERY slow at 1M cohort size.

Running the simulations in multiple threads will help a lot. And
deferred hashing in the SMT might help a little bit, too.